Add GitHub Actions build to replace CircleCI (#3)
### What
Add GitHub Actions build to replace CircleCI.

### Why
From time to time we have some painful error with CircleCI. Like just
now for some reason builds on this repository are failing. We were
experimenting with GitHub Actions on small projects, e.g. the
`stellar/go-xdr` project is using it. This is a small project, and it
was easier for me to copy of the GitHub Actions build file than it was
to debug CircleCI again.