fix handling versions that do not conform to semver (#2)
### What
When encountering a version that doesn't conform to the standard semver version format (vX.Y.Z) or the format that Go uses for tagless versions (v0.0.0-date-hash), fallback to just using the raw version name as the version.

### Why
While most developers use either semver tags for their versions, or no tags, some developers use non-standard formats for their tags and when we encounter these we should just use this as the version and not try to parse commit hashes out of them. The code is currently trying to parse a commit hash (the third component) out of `v0.0.1-2020.1.3` and raising an error because there is no third component.

This bug was discovered when using `golistcmp` on the `go list -m -json all` output for the dependencies in stellar/go#2430.